requirements.txt
alembic==1.18.3
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
fastapi==0.128.0
h11==0.16.0
idna==3.11
iniconfig==2.3.0
jinja2==3.1.4
macholib==1.16.4
mako==1.3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
packaging==26.0
pillow==12.1.0
pluggy==1.6.0
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
pytest==9.0.2
python-multipart==0.0.22
rarfile==4.2
rich==14.3.1
setuptools==80.10.2
shellingham==1.5.4
sqlalchemy==2.0.46
sqlmodel==0.0.31
starlette==0.50.0
typer==0.21.1
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.40.0
watchdog==6.0.0
